:truck: Rename test-data to samples
