+++ samples/short-and-sweet.html
